+++ services/vaults/holdings/save/handler.js
@@ -13,12 +13,33 @@ const delay = require('delay');
 const vaults = require('./vaults');
 const { delayTime } = require('./config');
 const axios = require('axios');
-const { getHoldings, getPoolTotalSupply } = require('./getHoldings');
+const { getHoldings } = require('./getHoldings');
 
 const db = dynamodb.doc;
 const Web3 = require('web3');
 
 const web3 = new Web3(process.env.WEB3_ENDPOINT);
+const infuraWeb3 = new Web3(web3);
+const getVirtualPrice = require('../../apy/save/handler');
+
+const pools = [
+  {
+    symbol: 'yCRV',
+    address: '0x45F783CCE6B7FF23B2ab2D70e416cdb7D6055f51',
+  },
+  {
+    symbol: 'crvBUSD',
+    address: '0x79a8C46DeA5aDa233ABaFFD40F3A0A2B1e5A4F27',
+  },
+  {
+    symbol: 'crvBTC',
+    address: '0x7fC77b5c7614E1533320Ea6DDc2Eb61fa00A9714',
+  },
+  {
+    symbol: 'ySUSD',
+    address: '0xF61718057901F84C4eEC4339EF8f0D86D2B45600',
+  }
+];
 
 const saveVault = async data => {
   const params = {
@@ -70,6 +91,29 @@ const readVault = async vault => {
   }
 };
 
+// EARN product
+const getPoolTotalSupply = async poolAddress => {
+  const poolMinABI = [
+    {
+      name: 'totalSupply',
+      outputs: [
+        {
+          type: 'uint256',
+          name: 'out',
+        },
+      ],
+      inputs: [],
+      constant: true,
+      payable: false,
+      type: 'function',
+      gas: 1181,
+    },
+  ];
+  const poolContract = new web3.eth.Contract(poolMinABI, poolAddress);
+  const _totalSupply = (await poolContract.methods.totalSupply().call()) / 1e18;
+  return _totalSupply;
+};
+
 // getting the veCRV locked in yearn used to vote on Curve
 const readveCRV = async () => {
   const veCRVMinABI = [
@@ -114,6 +158,28 @@ const readveCRV = async () => {
   return veCRVContract;
 };
 
+const getEarnHoldings = async (pool) => {
+  const { symbol, erc20address } = pool;
+  const currentBlockNbr = await infuraWeb3.eth.getBlockNumber();
+  await delay(delayTime);
+  const poolAddress = pool.address;
+  const virtualPriceCurrent =
+    (await getVirtualPrice.getVirtualPrice(poolAddress, currentBlockNbr)) /
+    1e18;
+  console.log(virtualPriceCurrent);
+  const poolTotalSupply = await getPoolTotalSupply(poolAddress);
+
+  poolBalance = poolTotalSupply * virtualPriceCurrent;
+  earnHoldings = {
+    tokenSymbol: symbol,
+    tokenAddress: erc20address,
+    poolBalanceUSD: poolBalance,
+  };
+  return earnHoldings;
+};
+
+
+
 // getting YFI staked in GOV for the holdings endpoint
 const readStaking = async () => {
   const staked = await getPoolTotalSupply(
@@ -137,6 +203,12 @@ const readStaking = async () => {
 
 module.exports.handler = async () => {
   const vaultsWithHoldings = [];
+      /*   const pool = _.find(pools, { symbol }); */
+
+      for(const pool of pools){
+        vaultsWithHoldings.push(await getEarnHoldings(pool));
+      }
+
   for (const vault of vaults) {
     const vaultWithHoldings = await readVault(vault);
     if (vaultWithHoldings !== null) {
@@ -144,6 +216,7 @@ module.exports.handler = async () => {
     }
     await delay(delayTime);
   }
+
 
   const staked = await readStaking();
   const veCRVLocked = await readveCRV();
